% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_demo_data.R
\name{gen_demo_data}
\alias{gen_demo_data}
\title{genenate a list of data with missing observation}
\usage{
gen_demo_data(
  N = 20000,
  R = 50,
  T = 5,
  K = 5,
  rho1 = -0.3,
  rho2 = 0.4,
  sigu2 = 0.6,
  zeta = 0.4
)
}
\arguments{
\item{rho1, rho2, sigu2, zeta:}{parameters you give to test}

\item{N:}{number of observations}

\item{R:}{repeating times}

\item{T:}{time periods}

\item{K:}{something you know}
}
\value{
a list:

list$b is a initial parameter vector;

list$data is a list: data.frame X, Y with missing data, K1, K2,T = T,D xi,R,N.

list$real_para is a vector of the real paramaters.
}
\description{
genenate a list of data with missing observation for demo tests
}
\examples{
input_data = gen_demo_data()
input_data = gen_demo_data(N=20000,R=50,T=5,K=5,rho1 = -0.3, rho2 = 0.4, sigu2 = 0.6,zeta = 0.4)
b = input_data$b
data_list = input_data$data_list
}
