% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenkins_main.R
\name{jenkins_main}
\alias{jenkins_main}
\title{the main function to estimate models}
\usage{
jenkins_main(b, data_list, real_para = NULL)
}
\arguments{
\item{b:}{a vector of initial parameters}

\item{data_list:}{a list of given data}

\item{real_para:}{a vector of real parameters you want to test}
}
\value{
result: list of coef of MLE, LIST OF PARAMETERS.
}
\description{
the main function to estimate models
}
\examples{
input_data <- gen_demo_data(N=20000,R=50,T=5,K=5,rho1 = -0.3, rho2 = 0.4, sigu2 = 0.6,zeta = 0.4)
b <- input_data$b
data_list <- input_data$data
real_para <- input_data$real_para
#main function to produce the result
jenkins_main(b=b, data_list = data_list, real_para = real_para)
}
